requirements.txt
crawl4ai==0.6.0
openai>=1.3.7
python-dotenv>=1.0.0
transformers>=4.32.0
torch>=2.0.0
tiktoken>=0.5.1
requests>=2.31.0
tqdm>=4.66.1
pyyaml>=6.0
